Refactoring CustomExceptionHandler classes (#1938)
