+++ services/businessconfig/src/main/java/org/opfab/businessconfig/controllers/CustomExceptionHandler.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2021, RTE (http://www.rte-france.com)
+/* Copyright (c) 2018-2022, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,11 +7,10 @@
  * This file is part of the OperatorFabric project.
  */
 
-
-
 package org.opfab.businessconfig.controllers;
 
 import lombok.extern.slf4j.Slf4j;
+import org.opfab.springtools.OpfabCustomExceptionHandler;
 import org.opfab.springtools.error.model.ApiError;
 import org.opfab.springtools.error.model.ApiErrorException;
 import org.springframework.http.HttpStatus;
@@ -19,7 +18,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -37,9 +35,7 @@ import java.io.IOException;
  */
 @RestControllerAdvice
 @Slf4j
-public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
-
-  public static final String GENERIC_MSG = "Caught exception at API level";
+public class CustomExceptionHandler extends OpfabCustomExceptionHandler {
 
   /**
    * Handles {@link IOException} as 404 errors
@@ -49,7 +45,7 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
    */
   @ExceptionHandler(IOException.class)
   public ResponseEntity<Object> handleIOException(IOException exception, final WebRequest request) {
-    log.warn(GENERIC_MSG,exception);
+    log.warn(GENERIC_MSG, request, exception);
     ApiError error = ApiError.builder()
        .status(HttpStatus.BAD_REQUEST)
        .message("Unable to load resource with specified request parameters")
@@ -67,26 +63,13 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
   @ExceptionHandler(FileNotFoundException.class)
   public ResponseEntity<Object> handleFileNotFoundException(FileNotFoundException exception, final WebRequest
      request) {
-    log.info(GENERIC_MSG,exception);
+    log.info(GENERIC_MSG, request, exception);
     ApiError error = ApiError.builder()
        .status(HttpStatus.NOT_FOUND)
        .message("The specified resource does not exist")
        .error(exception.getMessage())
        .build();
     return new ResponseEntity<>(error, error.getStatus());
-  }
-
-  /**
-   * Handles {@link ApiErrorException}
-   * @param exception exception to handle
-   * @param request Corresponding request of exchange
-   * @return Computed http response for specified exception
-   */
-  @ExceptionHandler(ApiErrorException.class)
-  public ResponseEntity<Object> handleApiError(ApiErrorException exception, final WebRequest
-     request) {
-    log.info(GENERIC_MSG,exception);
-    return new ResponseEntity<>(exception.getError(), exception.getError().getStatus());
   }
 
 }
