+++ services/users/src/main/java/org/opfab/users/controllers/CustomExceptionHandler.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2021, RTE (http://www.rte-france.com)
+/* Copyright (c) 2018-2022, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,23 +7,20 @@
  * This file is part of the OperatorFabric project.
  */
 
-
-
 package org.opfab.users.controllers;
 
 import feign.FeignException;
 import lombok.extern.slf4j.Slf4j;
+import org.opfab.springtools.OpfabCustomExceptionHandler;
 import org.opfab.springtools.error.model.ApiError;
 import org.opfab.springtools.error.model.ApiErrorException;
 import org.springframework.core.convert.ConversionFailedException;
 import org.springframework.dao.DuplicateKeyException;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 /**
  * CustomExceptionHandler.
@@ -39,22 +36,7 @@ import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExcep
  */
 @RestControllerAdvice
 @Slf4j
-public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
-
-  public static final String GENERIC_MSG = "Caught exception at API level";
-
-  /**
-   * Handles {@link ApiErrorException}
-   * @param exception exception to handle
-   * @param request Corresponding request of exchange
-   * @return Computed http response for specified exception
-   */
-  @ExceptionHandler(ApiErrorException.class)
-  public ResponseEntity<Object> handleApiError(ApiErrorException exception, final WebRequest
-     request) {
-    log.info(GENERIC_MSG + " : " + exception.getError().getMessage());
-    return new ResponseEntity<>(exception.getError(), exception.getError().getStatus());
-  }
+public class CustomExceptionHandler extends OpfabCustomExceptionHandler {
 
   /**
    * Handles {@link DuplicateKeyException} as 400 BAD_REQUEST error
@@ -64,7 +46,7 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
   @ExceptionHandler(DuplicateKeyException.class)
   public ResponseEntity<Object> handleDuplicateKey(DuplicateKeyException exception, final WebRequest
           request) {
-    log.error(GENERIC_MSG,exception);
+    log.error(GENERIC_MSG, request, exception);
     ApiError error = ApiError.builder()
             .status(HttpStatus.BAD_REQUEST)
             .message("Resource creation failed because a resource with the same key already exists.")
@@ -81,19 +63,13 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
   @ExceptionHandler(FeignException.class)
   public ResponseEntity<Object> handleFeignException(FeignException exception, final WebRequest
           request) {
-    log.error(GENERIC_MSG,exception);
+    log.error(GENERIC_MSG, request, exception);
     ApiError error = ApiError.builder()
             .status(HttpStatus.BAD_REQUEST)
             .message("Action failed because users privileges couldn't be retrieved from the Users service.")
             .error(exception.getMessage())
             .build();
     return new ResponseEntity<>(error, error.getStatus());
-  }
-
-  @Override
-  protected ResponseEntity<Object> handleExceptionInternal(Exception ex, Object body, HttpHeaders headers, HttpStatus status, WebRequest request) {
-    log.error("Uncaught internal server exception",ex);
-    return super.handleExceptionInternal(ex, body, headers, status, request);
   }
 
     /**
@@ -104,7 +80,7 @@ public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
   @ExceptionHandler(ConversionFailedException.class)
   public ResponseEntity<Object> handleConversionError(ConversionFailedException exception, final WebRequest
           request) {
-    log.error(GENERIC_MSG,exception);
+    log.error(GENERIC_MSG, request, exception);
     ApiError error = ApiError.builder()
             .status(HttpStatus.BAD_REQUEST)
             .message("Conversion Error")
