+++ tools/spring/spring-utilities/src/main/java/org/opfab/springtools/OpfabCustomExceptionHandler.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021-2022, RTE (http://www.rte-france.com)
+/* Copyright (c) 2022, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,10 +7,9 @@
  * This file is part of the OperatorFabric project.
  */
 
-package org.opfab.cards.publication.controllers;
+package org.opfab.springtools;
 
 import lombok.extern.slf4j.Slf4j;
-import org.opfab.springtools.OpfabCustomExceptionHandler;
 import org.opfab.springtools.error.model.ApiError;
 import org.opfab.springtools.error.model.ApiErrorException;
 import org.springframework.http.HttpHeaders;
@@ -19,13 +18,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
-
-import javax.validation.ConstraintViolationException;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 /**
- * CustomExceptionHandler.
+ * OpfabCustomExceptionHandler.
  * <ul>
- *     <li>Handle {@link ConstraintViolationException} as 400 BAD REQUEST errors</li>
  *     <li>Handle Api errors according to their configuration</li>
  *     <li>Handle uncaught logging errors</li>
  * </ul>
@@ -36,36 +33,27 @@ import javax.validation.ConstraintViolationException;
  */
 @RestControllerAdvice
 @Slf4j
-public class CustomExceptionHandler extends OpfabCustomExceptionHandler {
+public class OpfabCustomExceptionHandler extends ResponseEntityExceptionHandler {
+
+    public static final String GENERIC_MSG = "Caught exception at API level for request {}";
 
     /**
-     * Handles {@link ConstraintViolationException}
+     * Handles {@link ApiErrorException}
      * @param exception exception to handle
      * @param request Corresponding request of exchange
      * @return Computed http response for specified exception
      */
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity<Object> handleConstraintViolationException(ConstraintViolationException exception, final WebRequest
+    @ExceptionHandler(ApiErrorException.class)
+    public ResponseEntity<Object> handleApiError(ApiErrorException exception, final WebRequest
             request) {
         log.info(GENERIC_MSG, request, exception);
-        ApiError error = ApiError.builder()
-                .status(HttpStatus.BAD_REQUEST)
-                .message("Constraint violation in the request")
-                .error(exception.getMessage())
-                .build();
-        return new ResponseEntity<>(error, error.getStatus());
+        return new ResponseEntity<>(exception.getError(), exception.getError().getStatus());
     }
 
     @Override
     protected ResponseEntity<Object> handleExceptionInternal(Exception ex, Object body, HttpHeaders headers, HttpStatus status, WebRequest request) {
-        log.error("Uncaught internal server exception", ex);
-
-        ApiError error = ApiError.builder()
-                .status(status)
-                .message("Uncaught internal server exception")
-                .error(ex.getMessage())
-                .build();
-
-        return new ResponseEntity<>(error, error.getStatus());
+        log.error("Uncaught internal server exception",ex);
+        return super.handleExceptionInternal(ex, body, headers, status, request);
     }
 }
+
